Command center upgrade
It works! Almost a first playable. It’s even got the kill count
affecting spawn cool down, it’s pretty fun actually. Needs a lot of
tinkering but it’s going somewhere. Next steps I think is leaderboard
and we can wrap it up or tweak the gameplay more to heart’s content :)